koelsch: Resurrect support for Renesas M2 Koelsch board in auto board config
When the GDP Maintainers move from a 'branch per board' git rep
organisation to 'all boards in single master branch' they concentrated
on initially supporting a sub-set of the existing supported boards in
the mechanism to automatically configure a board. That list did not include
the Renesas R-Car M2 Koelsch Evaluation board. This commit resurrects
support for it in the new configuration mechanism.