r-car gen 3: Add support for Renesas R-Car M3 Starter Kit
Add the Renesas R-Car M3 Starter Kit [1] "low cost board"
to the board template system so it can be target board
for automatic build setup.

The Yocto BSP for the Starter Kit is not yet merged into
the upstream R-Car Gen 3 customer Yocto BSP. So add a
sub-module for the former. When it is merged we can
drop it and point at the upstream.

[1] http://www.elinux.org/R-Car/Boards/M3SK