git-submodule-bump: Script allows larger changelogs
The "shortlog" (change log) included in every layer update can get so
long that it can't fit into a shell variable.  Now the script uses a
temporary file instead.

It is no longer required to give a summary line since 99% of the time I
have found that the commit message will be edited anyway, so that can be
done in the editor.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>